add EquityPool
